+++ src/main/java/com/muhammaddaffa/nextcredits/transaction/TransactionLog.java
@@ -4,7 +4,7 @@ import java.util.UUID;
 
 public record TransactionLog(
         UUID playerId,
-        int credits,
+        double credits,
         String description,
         long timestamp
 ) {
